requirements.txt
alabaster==0.7.8
Babel==2.3.4
backports.shutil-get-terminal-size==1.0.0
coverage==4.0.3
decorator==4.0.9
Django==1.9.4
django-allauth==0.24.1
django-autoslug==1.9.3
django-braces==1.8.1
django-compat==1.0.11
django-countries==3.4.1
django-coverage-plugin==1.2.2
django-crispy-forms==1.6.0
django-debug-toolbar==1.4
django-environ==0.4.0
django-extensions==1.6.1
django-floppyforms==1.6.1
django-hijack==2.1.1
django-model-utils==2.4
django-redis==4.3.0
django-secure==1.0.1
django-test-plus==1.0.12
docutils==0.12
factory-boy==2.6.1
fake-factory==0.5.7
imagesize==0.7.1
ipaddress==1.0.16
ipdb==0.9.0
ipython==4.2.0
ipython-genutils==0.1.0
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.1.1
pathlib2==2.1.0
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.1.1
psycopg2==2.6.1
ptyprocess==0.5.1
py==1.4.31
Pygments==2.1.3
pytest==2.9.1
pytest-django==2.9.1
pytest-sugar==0.5.1
python-dateutil==2.5.3
python-openid==2.2.5
pytz==2015.7
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
simplegeneric==0.8.1
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.2
sqlparse==0.1.19
termcolor==1.1.0
traitlets==4.2.1
unicode-slugify==0.1.3
Unidecode==0.4.19
Werkzeug==0.11.4
whitenoise==2.0.6
